Update index.html
Supports bidirectional communication with Unix shell using stdin and stdout. When the button is pressed, a specified string is sent by the client to the server via websocket. If the string matches per the server side code, a specific bash command/script will be run via spawn/stdin, and output will be send back to the client and displayed in the "log_content" div (stdout).